requirements.txt
requests
imageio
pillow
